move SimulationSharedInputs, sync_sources to plugin submodule
